SOFTWARE-3359 Add Xrootd multiuser tests

+++ osgtest/tests/test_450_xrootd.py
@@ -19,10 +19,9 @@ class TestXrootd(osgunittest.OSGTestCase):
         if core.config['xrootd.gsi'] == "ON":
             core.skip_ok_unless_installed('globus-proxy-utils')
         self.skip_bad_unless(core.state['xrootd.started-server'] is True, 'Server not running')
-
+        temp_dir = "/tmp/vdttest"
         hostname = socket.getfqdn()
         if core.config['xrootd.gsi'] == "ON":
-            temp_dir = "/tmp/vdttest"
             if not os.path.exists(temp_dir):
                 os.mkdir(temp_dir)
                 user = pwd.getpwnam(core.options.username)
@@ -38,12 +37,21 @@ class TestXrootd(osgunittest.OSGTestCase):
         fail = core.diagnose('xrdcp copy, local to URL',
                              command, status, stdout, stderr)
         file_copied = os.path.exists(os.path.join(temp_dir, 'copied_file.txt'))
-        shutil.rmtree(temp_dir)
-
+        if core.config['xrootd.multiuser'] != "ON":
+            shutil.rmtree(temp_dir)
         self.assertEqual(status, 0, fail)
         self.assert_(file_copied, 'Copied file missing')
 
-    def test_02_xrdcp_server_to_local(self):
+    def test_02_xrootd_multiuser(self):
+        core.skip_ok_unless_installed('xrootd', 'xrootd-client', 'xrootd-multiuser', by_dependency=True)
+        temp_dir = "/tmp/vdttest"
+        if core.config['xrootd.multiuser'] == "ON":
+            file_path = os.path.join(temp_dir, 'copied_file.txt')
+            result_perm = core.check_file_ownership(file_path, core.options.username)
+            shutil.rmtree(temp_dir)
+            self.assertEqual(result_perm, True) 
+
+    def test_03_xrdcp_server_to_local(self):
         core.skip_ok_unless_installed('xrootd', 'xrootd-client', by_dependency=True)
         if core.config['xrootd.gsi'] == "ON":
             core.skip_ok_unless_installed('globus-proxy-utils')
@@ -72,7 +80,7 @@ class TestXrootd(osgunittest.OSGTestCase):
         self.assertEqual(status, 0, fail)
         self.assert_(file_copied, 'Copied file missing')
 
-    def test_03_xrootd_fuse(self):
+    def test_04_xrootd_fuse(self):
         # This tests xrootd-fuse using a mount in /mnt
         core.skip_ok_unless_installed('xrootd', 'xrootd-client', by_dependency=True)
         self.skip_ok_unless(os.path.exists("/mnt"), "/mnt did not exist")
